chore: extract interface & EditSelectionDrawer
