+++ packages/shared/src/lib/auxiliary/wallet-connect/interface/index.ts
@@ -1,3 +1,4 @@
 export * from './connected-dapp.interface'
 export * from './dapp-metadata.interface'
 export * from './pairing.interface'
+export * from './selections.interface'
