+++ packages/shared/src/lib/auxiliary/wallet-connect/interface/selections.interface.ts
@@ -1,0 +1,7 @@
+import { IAccountState } from '@core/account/interfaces'
+
+export interface ISelections {
+    chains?: string[]
+    methods?: string[]
+    accounts?: IAccountState[]
+}
